+++ packages/web3-eth-accounts/src/index.ts
@@ -41,3 +41,5 @@ export * from './wallet';
 export * from './account';
 export * from './types';
 export * from './schemas';
+export * from './common';
+export * from './tx';